chore: move files from build to lib for better import, fix error in cdn
