+++ .eslintignore
@@ -2,4 +2,7 @@ coverage
 dist
 node_modules
 example
+# ignore for no typescript checking
 src
+# ignore for dont check build result
+lib
