+++ __tests__/domain.test.js
@@ -1,4 +1,4 @@
-const { Domain } = require('../build');
+const { Domain } = require('../lib');
 
 describe('Domain', () => {
   const credentials = {
