+++ package.json
@@ -2,7 +2,7 @@
   "name": "tencent-component-toolkit",
   "version": "1.20.10",
   "description": "Tencent component toolkit",
-  "main": "build/index.js",
+  "main": "lib/index.js",
   "scripts": {
     "build": "tsc -p .",
     "test": "jest",
@@ -76,7 +76,7 @@
     "lint-staged": "^10.0.8",
     "prettier": "^1.19.1",
     "semantic-release": "^17.0.4",
-    "typescript": "~4.1.0"
+    "typescript": "^4.1.0"
   },
   "dependencies": {
     "@tencent-sdk/capi": "^1.1.8",
